[move] PlatformType enum 패키지 이동 (#308)
- notification 패키지 하위로 이동
- 프로젝트 전체의 util 적인 성질보다는 notification 패키지 하위에서 사용되는 value 로 생각하는 것으로 결정됨
