+++ src/main/java/konkuk/thip/notification/adapter/in/web/request/FcmTokenRegisterRequest.java
@@ -3,7 +3,7 @@ package konkuk.thip.notification.adapter.in.web.request;
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
 import konkuk.thip.notification.application.port.in.dto.FcmTokenRegisterCommand;
-import konkuk.thip.common.util.PlatformType;
+import konkuk.thip.notification.domain.value.PlatformType;
 
 @Schema(description = "FCM 토큰 등록 요청 DTO")
 public record FcmTokenRegisterRequest(
