+++ src/main/java/konkuk/thip/notification/application/port/in/dto/FcmTokenRegisterCommand.java
@@ -1,6 +1,6 @@
 package konkuk.thip.notification.application.port.in.dto;
 
-import konkuk.thip.common.util.PlatformType;
+import konkuk.thip.notification.domain.value.PlatformType;
 import lombok.Builder;
 
 @Builder
